requirements.txt
autopep8==1.4.4
beautifulsoup4==4.8.2
bootstrap4==0.1.0
confusable-homoglyphs==3.2.0
Django==1.11
django-bootstrap3==12.0.3
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
django-forms-bootstrap==3.1.0
django-registration==2.4.1
django-tinymce==2.8.0
djangorestframework==3.11.0
Pillow==7.0.0
psycopg2==2.7.5
pycodestyle==2.5.0
python-decouple==3.3
pytz==2019.3
soupsieve==1.9.5
